refactor: minor reorganization
