+++ src/modules/project/pages/projectView/views/projectActivityPanel/screens/info/views/ActivityBrief.tsx
@@ -16,22 +16,22 @@ import { useCallback, useEffect, useMemo, useState } from 'react'
 import { useTranslation } from 'react-i18next'
 import { AiOutlineEllipsis } from 'react-icons/ai'
 
-import { SkeletonLayout } from '../../../../../../../../components/layouts'
-import { UserAvatar } from '../../../../../../../../components/ui/UserAvatar'
-import { useProjectContext } from '../../../../../../../../context'
-import { ExternalAccountType } from '../../../../../../../../pages/auth'
-import { useFollowedProjectsValue } from '../../../../../../../../pages/auth/state'
+import { SkeletonLayout } from '../../../../../../../../../components/layouts'
+import { UserAvatar } from '../../../../../../../../../components/ui/UserAvatar'
+import { useProjectContext } from '../../../../../../../../../context'
+import { ExternalAccountType } from '../../../../../../../../../pages/auth'
+import { useFollowedProjectsValue } from '../../../../../../../../../pages/auth/state'
 import {
   FunderWithUserFragment,
   OrderByOptions,
   ProjectMilestone,
   useProjectFundersQuery,
-} from '../../../../../../../../types'
-import { toInt, useMobileMode, useNotification } from '../../../../../../../../utils'
-import { getProjectBalance } from '../../../../../../../../utils/helpers'
-import { ContributeButton, FollowButton, ShareButton } from '../../../projectMainBody/components'
-import { BalanceDisplayButton, SubscribeButton } from './components'
-import { ProjectFundersModal, useProjectFundersModal } from './components/ProjectFundersModal'
+} from '../../../../../../../../../types'
+import { toInt, useMobileMode, useNotification } from '../../../../../../../../../utils'
+import { getProjectBalance } from '../../../../../../../../../utils/helpers'
+import { ContributeButton, FollowButton, ShareButton } from '../../../../projectMainBody/components'
+import { BalanceDisplayButton, SubscribeButton } from '../components'
+import { ProjectFundersModal, useProjectFundersModal } from '../components/ProjectFundersModal'
 
 const TIME_AFTER_WHICH_TOOLTIP_SHOULD_CLOSE_MILLIS = 1500
